dtos moved to presentation layer
